+++ minhas-financas-back-end/MinhasFinancas.Domain/Services/AnaliseFinanceira/Enums/CodigoPilarMfScoreFinanceiro.cs
@@ -1,0 +1,11 @@
+namespace MinhasFinancas.Domain.Services.AnaliseFinanceira.Enums
+{
+    public enum CodigoPilarMfScoreFinanceiro
+    {
+        FluxoDeCaixa = 0,
+        LiquidezEReserva = 1,
+        EndividamentoEObrigacoes = 2,
+        Patrimonio = 3,
+        PlanejamentoEDisciplina = 4
+    }
+}