+++ minhas-financas-back-end/MinhasFinancas.Domain/Services/AnaliseFinanceira/Enums/DirecaoTendenciaMfScoreFinanceiro.cs
@@ -1,0 +1,10 @@
+namespace MinhasFinancas.Domain.Services.AnaliseFinanceira.Enums
+{
+    public enum DirecaoTendenciaMfScoreFinanceiro
+    {
+        Positiva = 0,
+        Neutra = 1,
+        Negativa = 2,
+        Indeterminada = 3
+    }
+}